Giocatore.java creata e implementata, correzione errori
