+++ src/Giocatore.java
@@ -1,0 +1,20 @@
+
+public class Giocatore {
+	private static final int CFU_INIZIALI = 20;
+	
+	//private Borsa borsa;
+	private int cfu;
+	
+	public Giocatore() {
+		//this.borsa = new Borsa();
+		this.cfu= CFU_INIZIALI;
+	}
+	
+	public int getCfu() {
+		return this.cfu;
+	}
+
+	public void setCfu(int cfu) {
+		this.cfu = cfu;		
+	}
+}
