+++ src/Partita.java
@@ -10,17 +10,15 @@
  */
 
 public class Partita {
-
-	private static final int CFU_INIZIALI = 20;
 	
 	private Labirinto labirinto;
+	private Giocatore giocatore;
 	private boolean finita;
-	private int cfu;
 	
 	public Partita(){
 		this.labirinto = new Labirinto();
+		this.giocatore = new Giocatore();
 		this.finita = false;
-		this.cfu = CFU_INIZIALI;
 	}
 
     /**
@@ -45,7 +43,7 @@ public class Partita {
 	 * @return vero se partita finita
 	 */
 	public boolean isFinita() {
-		return finita || vinta() || (cfu == 0);
+		return finita || vinta() || (getGiocatore().getCfu() == 0);
 	}
 	
 	/**
@@ -60,11 +58,7 @@ public class Partita {
 		return this.labirinto;
 	}
 	
-	public int getCfu() {
-		return this.cfu;
+	public Giocatore getGiocatore() {
+		return this.giocatore;
 	}
-
-	public void setCfu(int cfu) {
-		this.cfu = cfu;		
-	}	
 }
